requirements.txt
python-dotenv
tabulate
openai>=1.0.0
